fix(frontend): 临时密码修改优化 #8250 # Reviewed, transaction id: 26128
