+++ dbm-ui/frontend/src/locales/zh-cn.json
@@ -3675,6 +3675,8 @@
   "冷/热节点": "冷/热节点",
   "冷_热节点规格": "冷_热节点规格",
   "不允许使用特殊账号名称n": "不允许使用特殊账号名称 {n}",
+  "密码不符合要求": "密码不符合要求",
+  "错误日志": "错误日志",
   "这行勿动！新增翻译请在上一行添加！": ""
 
 }
